feat: интерфейсы экшенов и union-types вынесены в отдельные файлы
