+++ src/services/actions/burger-ingredients.ts
@@ -7,28 +7,11 @@ import { SET_ACTIVE_TAB,
 import { TIngredient } from "../types/data";
 import { AppDispatch, AppThunk } from "../types";
 
-// Типизация экшенов
-export interface ISetActiveTabAction {
-  readonly type: typeof SET_ACTIVE_TAB;
-  readonly current: string;
-}
-export interface IGetIngredientsRequestAction {
-  readonly type: typeof GET_INGREDIENTS_REQUEST;
-}
-export interface IGetIngredientsFailedAction {
-  readonly type: typeof GET_INGREDIENTS_FAILED;
-}
-export interface IGetIngredientsSuccessAction {
-  readonly type: typeof GET_INGREDIENTS_SUCCESS;
-  readonly ingredients: ReadonlyArray<TIngredient>;
-}
-
-// Объединяем в Union
-export type TBurgerIngredientsActions =
-| ISetActiveTabAction
-| IGetIngredientsRequestAction
-| IGetIngredientsFailedAction
-| IGetIngredientsSuccessAction;
+import { ISetActiveTabAction,
+  IGetIngredientsRequestAction,
+  IGetIngredientsFailedAction,
+  IGetIngredientsSuccessAction
+} from '../interfaces/burger-ingredients-interfaces';
 
 export const setActiveTab = (value: string): ISetActiveTabAction => ({ type: SET_ACTIVE_TAB, current: value});
