+++ src/services/actions/ws-auth.ts
@@ -5,45 +5,13 @@ import { WS_AUTH_CONNECTION_START,
   WS_AUTH_GET_MESSAGE,
   WS_AUTH_SEND_MESSAGE } from "../action-types/ws-auth-action-types";
 import { TOrder } from "../types/data";
-
-// Типизация экшенов
-export interface IWsAuthConnectionStartAction {
-  readonly type: typeof WS_AUTH_CONNECTION_START;
-  readonly payload: string;
-}
-export interface IWsAuthConnectionSuccessAction {
-  readonly type: typeof WS_AUTH_CONNECTION_SUCCESS;
-  readonly payload: string | undefined;
-}
-export interface IWsAuthConnectionErrorAction {
-  readonly type: typeof WS_AUTH_CONNECTION_ERROR;
-  readonly payload: string | undefined;
-}
-export interface IWsAuthConnectionClosedAction {
-  readonly type: typeof WS_AUTH_CONNECTION_CLOSED;
-  readonly payload: string | undefined;
-}
-export interface IWsAuthGetMessageAction {
-  readonly type: typeof WS_AUTH_GET_MESSAGE;
-  readonly payload: { 
-    readonly error: string | undefined;
-    readonly orders: ReadonlyArray<TOrder>;
-    readonly total: number;
-    readonly totalToday: number;
-  }
-}
-export interface IWsAuthSendMessageAction {
-  readonly type: typeof WS_AUTH_SEND_MESSAGE;
-}
-
-// Объединяем в Union
-export type TWsAuthActions =
-| IWsAuthConnectionStartAction
-| IWsAuthConnectionSuccessAction
-| IWsAuthConnectionErrorAction
-| IWsAuthConnectionClosedAction
-| IWsAuthGetMessageAction
-| IWsAuthSendMessageAction;
+import { IWsAuthConnectionStartAction,
+  IWsAuthConnectionSuccessAction,
+  IWsAuthConnectionErrorAction,
+  IWsAuthConnectionClosedAction,
+  IWsAuthGetMessageAction,
+  IWsAuthSendMessageAction
+} from '../interfaces/ws-auth-interfaces';
 
 export const wsAuthConnectionStartAction = (url: string): IWsAuthConnectionStartAction => ({ type: WS_AUTH_CONNECTION_START, payload: url });
 export const wsAuthConnectionSuccessAction = (error: string | undefined): IWsAuthConnectionSuccessAction => ({ type: WS_AUTH_CONNECTION_SUCCESS, payload: error })
