+++ src/services/actions/ws.ts
@@ -5,45 +5,13 @@ import { WS_CONNECTION_START,
   WS_GET_MESSAGE,
   WS_SEND_MESSAGE } from "../action-types/ws-action-types";
 import { TOrder } from "../types/data";
-
-// Типизация экшенов
-export interface IWsConnectionStartAction {
-  readonly type: typeof WS_CONNECTION_START;
-  readonly payload: string;
-}
-export interface IWsConnectionSuccessAction {
-  readonly type: typeof WS_CONNECTION_SUCCESS;
-  readonly payload: string | undefined;
-}
-export interface IWsConnectionErrorAction {
-  readonly type: typeof WS_CONNECTION_ERROR;
-  readonly payload: string | undefined;
-}
-export interface IWsConnectionClosedAction {
-  readonly type: typeof WS_CONNECTION_CLOSED;
-  readonly payload: string | undefined;
-}
-export interface IWsGetMessageAction {
-  readonly type: typeof WS_GET_MESSAGE;
-  readonly payload: { 
-    readonly error: string | undefined;
-    readonly orders: ReadonlyArray<TOrder>;
-    readonly total: number;
-    readonly totalToday: number;
-  }
-}
-export interface IWsSendMessageAction {
-  readonly type: typeof WS_SEND_MESSAGE;
-}
-
-// Объединяем в Union
-export type TWsActions =
-| IWsConnectionStartAction
-| IWsConnectionSuccessAction
-| IWsConnectionErrorAction
-| IWsConnectionClosedAction
-| IWsGetMessageAction
-| IWsSendMessageAction;
+import { IWsConnectionStartAction,
+  IWsConnectionSuccessAction,
+  IWsConnectionErrorAction,
+  IWsConnectionClosedAction,
+  IWsGetMessageAction,
+  IWsSendMessageAction
+} from '../interfaces/ws-interfaces';
 
 export const wsConnectionStartAction = (url: string): IWsConnectionStartAction => ({ type: WS_CONNECTION_START, payload: url });
 export const wsConnectionSuccessAction = (error: string | undefined): IWsConnectionSuccessAction => ({ type: WS_CONNECTION_SUCCESS, payload: error })
