+++ src/services/reducers/order-details.ts
@@ -2,7 +2,7 @@ import { POST_ORDER_REQUEST,
     POST_ORDER_FAILED,
     POST_ORDER_SUCCESS,
     CLOSE_ORDER_DETAILS_MODAL } from "../action-types/order-details-action-types";
-import { TOrderDetailsActions } from "../actions/order-details";
+import { TOrderDetailsActions } from "../action-types/union-types";
 
 type TOrderDetailsInitialState = {
     orderRequest: boolean;
